+++ src/controllers/VerifyController.php
@@ -9,6 +9,9 @@ use craft\helpers\UrlHelper;
 use born05\twofactorauthentication\Plugin as TwoFactorAuth;
 use born05\twofactorauthentication\web\assets\verify\VerifyAsset;
 
+use yii\base\Event;
+use yii\web\UserEvent;
+
 class VerifyController extends Controller
 {
     /**
@@ -28,7 +31,6 @@ class VerifyController extends Controller
     {
         $this->requireLogin();
         $this->requirePostRequest();
-        $responseService = TwoFactorAuth::$plugin->response;
         $request = Craft::$app->getRequest();
 
         $authenticationCode = $request->getBodyParam('authenticationCode');
@@ -37,6 +39,12 @@ class VerifyController extends Controller
         $user = Craft::$app->getUser()->getIdentity();
 
         if (TwoFactorAuth::$plugin->verify->verify($user, $authenticationCode)) {
+            // Get the session duration
+            $sessionDuration = Craft::$app->getUser()->getRemainingSessionTime();
+
+            // Throw the after login event, because we blocked it earlier for non-cookieBased events.
+            $this->afterLogin($user, false, $sessionDuration);
+
             return $this->_handleSuccessfulLogin(true);
         } else {
             $errorCode = User::AUTH_INVALID_CREDENTIALS;
@@ -71,23 +79,49 @@ class VerifyController extends Controller
     private function _handleSuccessfulLogin($setNotice)
     {
         $request = Craft::$app->getRequest();
-        $responseService = TwoFactorAuth::$plugin->response;
-        $returnUrl = $responseService->getReturnUrl();
+        $returnUrl = TwoFactorAuth::$plugin->response->getReturnUrl();
 
         // If this was an Ajax request, just return success:true
         if ($request->getAcceptsJson()) {
-            return $this->asJson([
+            $return = [
                 'success' => true,
-                'returnUrl' => $returnUrl
-            ]);
+                'returnUrl' => $returnUrl,
+            ];
+
+            if (Craft::$app->getConfig()->getGeneral()->enableCsrfProtection) {
+                $return['csrfTokenValue'] = $request->getCsrfToken();
+            }
+
+            return $this->asJson($return);
         }
-        
+
         if ($setNotice) {
-            Craft::$app->getSession()->setNotice(Craft::t('two-factor-authentication', 'Logged in.'));
+            Craft::$app->getSession()->setNotice(Craft::t('app', 'Logged in.'));
         }
 
         $user = Craft::$app->getUser()->getIdentity();
 
         return $this->redirectToPostedUrl($user, $returnUrl);
+    }
+
+    /**
+     * COPIED FROM \yii\web\User
+     *
+     * This method is called after the user is successfully logged in.
+     * The default implementation will trigger the [[EVENT_AFTER_LOGIN]] event.
+     * If you override this method, make sure you call the parent implementation
+     * so that the event is triggered.
+     * @param IdentityInterface $identity the user identity information
+     * @param bool $cookieBased whether the login is cookie-based
+     * @param int $duration number of seconds that the user can remain in logged-in status.
+     * If 0, it means login till the user closes the browser or the session is manually destroyed.
+     */
+    protected function afterLogin($identity, $cookieBased, $duration)
+    {
+        Event::trigger(\craft\web\User::class, \craft\web\User::EVENT_AFTER_LOGIN, new UserEvent([
+            'identity' => $identity,
+            'cookieBased' => $cookieBased,
+            'duration' => $duration,
+        ]));
     }
 }