+++ src/services/Response.php
@@ -26,7 +26,14 @@ class Response extends Component
         // Get the return URL
         $userService = Craft::$app->getUser();
         $request = Craft::$app->getRequest();
-        $returnUrl = $userService->getReturnUrl();
+
+        if (!Craft::$app->getRequest()->getIsCpRequest()) {
+            $returnUrl = Craft::$app->getRequest()->getValidatedBodyParam('redirect');
+        }
+
+        if (!isset($returnUrl) || is_null($returnUrl)) {
+            $returnUrl = $userService->getReturnUrl();
+        }
 
         // Clear it out
         $userService->removeReturnUrl();
@@ -38,9 +45,9 @@ class Response extends Component
             TwoFactorAuth::$plugin->request->is2FASpecialRequests()
         ) {
             // Is this a CP request and can they access the CP?
-            if (Craft::$app->getRequest()->getIsCpRequest() && Craft::$app->getUser()->checkPermission('accessCp')) {
+            if (Craft::$app->getRequest()->getIsCpRequest()) {
                 $returnUrl = UrlHelper::cpUrl(Craft::$app->getConfig()->getGeneral()->getPostCpLoginRedirect());
-            } else {
+            } elseif (is_null($returnUrl)) {
                 $returnUrl = UrlHelper::siteUrl(Craft::$app->getConfig()->getGeneral()->getPostLoginRedirect());
             }
         }